Move mapping file creation to new step
The step is only added to the test case if SLM is included in the run
